+++ sickbeard/__init__.py
@@ -421,11 +421,12 @@ PUSHBULLET_DEVICE_IDEN = None
 USE_SLACK = False
 SLACK_NOTIFY_ONSNATCH = False
 SLACK_NOTIFY_ONDOWNLOAD = False
-SLACK_ACCESS_TOKEN = None
+SLACK_NOTIFY_ONSUBTITLEDOWNLOAD = False
 SLACK_CHANNEL = None
 SLACK_AS_USER = False
 SLACK_BOT_NAME = None
 SLACK_ICON_URL = None
+SLACK_ACCESS_TOKEN = None
 
 USE_EMAIL = False
 EMAIL_OLD_SUBJECTS = None
@@ -640,7 +641,8 @@ def initialize(console_logging=True):
             USE_TRAKT, TRAKT_CONNECTED_ACCOUNT, TRAKT_ACCOUNTS, TRAKT_MRU, TRAKT_VERIFY, \
             TRAKT_USE_WATCHLIST, TRAKT_REMOVE_WATCHLIST, TRAKT_TIMEOUT, TRAKT_METHOD_ADD, TRAKT_START_PAUSED, \
             TRAKT_SYNC, TRAKT_DEFAULT_INDEXER, TRAKT_REMOVE_SERIESLIST, TRAKT_UPDATE_COLLECTION, \
-            USE_SLACK, SLACK_NOTIFY_ONSNATCH, SLACK_NOTIFY_ONDOWNLOAD, SLACK_ACCESS_TOKEN, SLACK_CHANNEL, SLACK_AS_USER, SLACK_BOT_NAME, SLACK_ICON_URL, \
+            USE_SLACK, SLACK_NOTIFY_ONSNATCH, SLACK_NOTIFY_ONDOWNLOAD, SLACK_NOTIFY_ONSUBTITLEDOWNLOAD, \
+            SLACK_CHANNEL, SLACK_AS_USER, SLACK_BOT_NAME, SLACK_ICON_URL, SLACK_ACCESS_TOKEN, \
             USE_EMAIL, EMAIL_NOTIFY_ONSNATCH, EMAIL_NOTIFY_ONDOWNLOAD, EMAIL_NOTIFY_ONSUBTITLEDOWNLOAD, EMAIL_FROM, \
             EMAIL_HOST, EMAIL_PORT, EMAIL_TLS, EMAIL_USER, EMAIL_PASSWORD, EMAIL_LIST, EMAIL_OLD_SUBJECTS
         # Anime Settings
@@ -1046,11 +1048,12 @@ def initialize(console_logging=True):
         USE_SLACK = bool(check_setting_int(CFG, 'Slack', 'use_slack', 0))
         SLACK_NOTIFY_ONSNATCH = bool(check_setting_int(CFG, 'Slack', 'slack_notify_onsnatch', 0))
         SLACK_NOTIFY_ONDOWNLOAD = bool(check_setting_int(CFG, 'Slack', 'slack_notify_ondownload', 0))
-        SLACK_ACCESS_TOKEN = check_setting_str(CFG, 'Slack', 'slack_access_token', '')
+        SLACK_NOTIFY_ONSUBTITLEDOWNLOAD = bool(check_setting_int(CFG, 'Slack', 'slack_notify_onsubtitledownload', 0))
         SLACK_CHANNEL = check_setting_str(CFG, 'Slack', 'slack_channel', '')
         SLACK_AS_USER = bool(check_setting_int(CFG, 'Slack', 'slack_as_user', 0))
         SLACK_BOT_NAME = check_setting_str(CFG, 'Slack', 'slack_bot_name', '')
         SLACK_ICON_URL = check_setting_str(CFG, 'Slack', 'slack_icon_url', '')
+        SLACK_ACCESS_TOKEN = check_setting_str(CFG, 'Slack', 'slack_access_token', '')
 
         USE_EMAIL = bool(check_setting_int(CFG, 'Email', 'use_email', 0))
         EMAIL_OLD_SUBJECTS = bool(check_setting_int(CFG, 'Email', 'email_old_subjects',
@@ -1844,11 +1847,12 @@ def save_config():
     new_config['Slack']['use_slack'] = int(USE_SLACK)
     new_config['Slack']['slack_notify_onsnatch'] = int(SLACK_NOTIFY_ONSNATCH)
     new_config['Slack']['slack_notify_ondownload'] = int(SLACK_NOTIFY_ONDOWNLOAD)
-    new_config['Slack']['slack_access_token'] = SLACK_ACCESS_TOKEN
+    new_config['Slack']['slack_notify_onsubtitledownload'] = int(SLACK_NOTIFY_ONSUBTITLEDOWNLOAD)
     new_config['Slack']['slack_channel'] = SLACK_CHANNEL
     new_config['Slack']['slack_as_user'] = int(SLACK_AS_USER)
     new_config['Slack']['slack_bot_name'] = SLACK_BOT_NAME
     new_config['Slack']['slack_icon_url'] = SLACK_ICON_URL
+    new_config['Slack']['slack_access_token'] = SLACK_ACCESS_TOKEN
 
     new_config['Email'] = {}
     new_config['Email']['use_email'] = int(USE_EMAIL)