+++ sickbeard/notifiers/slack.py
@@ -2,7 +2,7 @@
 #
 # This file is part of SickGear.
 #
-# Thanks to: mallen86
+# Thanks to: mallen86, generica
 #
 # SickGear is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -17,74 +17,50 @@
 # You should have received a copy of the GNU General Public License
 # along with SickGear.  If not, see <http://www.gnu.org/licenses/>.
 
-import json
 import sickbeard
+from sickbeard import common, logger
 
-from sickbeard import logger, common
 
 class SlackNotifier:
+    def __init__(self):
+        self.sg_logo_url = 'https://raw.githubusercontent.com/SickGear/SickGear/master' + \
+                           '/gui/slick/images/ico/apple-touch-icon-precomposed.png'
 
-    def _send_to_slack(self, message, accessToken, channel, as_user, bot_name, icon_url):
-        SLACK_ENDPOINT = "https://slack.com/api/chat.postMessage"
+    def _notify(self, msg, channel='', as_user=False, bot_name='', icon_url='', access_token='', force=False):
+        custom = (force and not as_user) or (not (force or sickbeard.SLACK_AS_USER))
+        resp = (sickbeard.USE_SLACK or force) and sickbeard.helpers.getURL(
+            url='https://slack.com/api/chat.postMessage',
+            post_data=dict(
+                [('text', msg), ('token', (access_token, sickbeard.SLACK_ACCESS_TOKEN)[not access_token]),
+                 ('channel', (channel, sickbeard.SLACK_CHANNEL)[not channel]), ('as_user', not custom)] +
+                ([], [('username', (bot_name, sickbeard.SLACK_BOT_NAME or 'SickGear')[not bot_name]),
+                      ('icon_url', (icon_url, sickbeard.SLACK_ICON_URL or self.sg_logo_url)[not icon_url])])[custom]),
+            json=True)
 
-        data = {}
-        data["token"] = accessToken
-        data["channel"] = channel
-        data["username"] = bot_name
-        data["text"] = message
-        data["icon_url"] = icon_url
-        data["as_user"] = as_user
+        result = resp and resp['ok'] or resp['error']
+        if True is not result:
+            logger.log(u'Slack failed sending message, response: "%s"' % resp['error'], logger.ERROR)
+        return result
 
-        urlResp = sickbeard.helpers.getURL(url=SLACK_ENDPOINT, post_data=data)
-        if urlResp:
-            resp = json.loads(urlResp)
-        else:
-            return False
+    def _notify_str(self, pre_text, post_text):
+        return self._notify('%s: %s' % (common.notifyStrings[pre_text].strip('#: '), post_text))
 
-        # if ("error" in resp):
-        #     raise Exception(resp["error"])
-
-        if (resp["ok"] == True):
-            logger.log(u"Slack: Succeeded sending message.", logger.MESSAGE)
-            return True
-
-        logger.log(u"Slack: Failed sending message: " + resp["error"], logger.ERROR)
-        return False
-
-    def _notify(self, message, accessToken='', channel='', as_user='', bot_name='', icon_url='', force=False):
-        # suppress notifications if the notifier is disabled but the notify options are checked
-        if not sickbeard.USE_SLACK and not force:
-            return False
-
-        if not accessToken:
-            accessToken = sickbeard.SLACK_ACCESS_TOKEN
-        if not channel:
-            channel = sickbeard.SLACK_CHANNEL
-        if not as_user:
-            as_user = sickbeard.SLACK_AS_USER
-        if not bot_name:
-            bot_name = sickbeard.SLACK_BOT_NAME
-        if not icon_url:
-            icon_url = sickbeard.SLACK_ICON_URL
-
-        return self._send_to_slack(message, accessToken, channel, as_user, bot_name, icon_url)
-
-##############################################################################
-# Public functions
-##############################################################################
+    def test_notify(self, channel, as_user, bot_name, icon_url, access_token):
+        return self._notify('This is a test notification from SickGear',
+                            channel, as_user, bot_name, icon_url, access_token, force=True)
 
     def notify_snatch(self, ep_name):
-        if sickbeard.SLACK_NOTIFY_ONSNATCH:
-            self._notify(common.notifyStrings[common.NOTIFY_SNATCH] + ': ' + ep_name)
+        return sickbeard.SLACK_NOTIFY_ONSNATCH and self._notify_str(common.NOTIFY_SNATCH, ep_name)
 
     def notify_download(self, ep_name):
-        if sickbeard.SLACK_NOTIFY_ONDOWNLOAD:
-            self._notify(common.notifyStrings[common.NOTIFY_DOWNLOAD] + ': ' + ep_name)
+        return sickbeard.SLACK_NOTIFY_ONDOWNLOAD and self._notify_str(common.NOTIFY_DOWNLOAD, ep_name)
 
-    def test_notify(self, accessToken, channel, as_user, bot_name, icon_url):
-        return self._notify("This is a test notification from SickGear", accessToken, channel, as_user, bot_name, icon_url, force=True)
+    def notify_subtitle_download(self, ep_name, lang):
+        return sickbeard.SLACK_NOTIFY_ONSUBTITLEDOWNLOAD and \
+               self._notify_str(common.NOTIFY_SUBTITLE_DOWNLOAD, '%s: %s' % (ep_name, lang))
 
-    def update_library(self, ep_obj):
-        pass
+    def notify_git_update(self, new_version='??'):
+        return self._notify_str(common.NOTIFY_GIT_UPDATE_TEXT, new_version)
+
 
 notifier = SlackNotifier